% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_qoi.R
\name{compute_qoi}
\alias{compute_qoi}
\title{Compute Quantities of Interest (GATE, GATEcv, URATE)}
\usage{
compute_qoi(fit_obj, algorithms)
}
\arguments{
\item{fit_obj}{An output object of \code{estimate_itr} function from the \code{evalITR} package.}

\item{algorithms}{Machine learning algorithms}
}
\description{
Compute Quantities of Interest (GATE, GATEcv, URATE)
}
