% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consist.test.R
\name{consist.test}
\alias{consist.test}
\title{The Consistency Test for Grouped Average Treatment Effects (GATEs) in Randomized Experiments}
\usage{
consist.test(D, tau, Y, ngates = 5, nsim = 10000, centered = TRUE)
}
\arguments{
\item{D}{Binary treatment indicator (0 or 1).}

\item{tau}{Continuous score vector.}

\item{Y}{Outcome vector.}

\item{ngates}{Number of groups (at least 2).}

\item{nsim}{Number of Monte Carlo draws (at least 2).}

\item{centered}{Whether to center outcomes before estimation.}
}
\value{
A list with test statistic \code{stat} and p-value \code{pval}.
}
\description{
Tests whether treatment effects are nondecreasing across score groups.
}
\details{
Uses \code{\link[evalITR]{consist.test}} for inference.
}
\examples{
D = c(1,0,1,0,1,0,1,0)
tau = c(0,0.1,0.2,0.3,0.4,0.5,0.6,0.7)
Y = c(4,5,0,2,4,1,-4,3)
consisttestlist <- consist.test(D,tau,Y,ngates=2)
consisttestlist$stat
consisttestlist$pval
}
\references{
Imai and Li (2022). \dQuote{Statistical Inference for Heterogeneous Treatment Effects Discovered by Generic Machine Learning in Randomized Experiments},
}
\author{
Michael Lingzhi Li, Technology and Operations Management, Harvard Business School
\email{mili@hbs.edu}, \url{https://www.michaellz.com/};
}
\keyword{evaluation}
