% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het.test.R
\name{het.test}
\alias{het.test}
\title{The Heterogeneity Test for Grouped Average Treatment Effects (GATEs) in Randomized Experiments}
\usage{
het.test(
  T,
  tau,
  Y,
  ngates = 5,
  center = TRUE,
  residualize = FALSE,
  X = NULL,
  p = mean(T),
  n_folds = 5,
  nuisance_learner = "SuperLearner",
  SL.library = c("SL.mean", "SL.glm", "SL.gam", "SL.glmnet", "SL.nnet", "SL.earth",
    "SL.rpartPrune"),
  SL_folds = 5,
  learner_args = list()
)
}
\arguments{
\item{T}{A vector of the unit-level binary treatment receipt variable for each sample.}

\item{tau}{A vector of the unit-level continuous score. Conditional Average Treatment Effect is one possible measure.}

\item{Y}{A vector of the outcome variable of interest for each sample.}

\item{ngates}{The number of groups to separate the data into. The groups are determined by \code{tau}. Default is 5.}

\item{center}{Whether to center the outcome before constructing the test
statistic. Default is \code{TRUE}.}

\item{residualize}{Whether to residualize \code{Y} before testing. Default is
\code{FALSE}, which uses the supplied outcome without nuisance adjustment.}

\item{X}{Optional covariate matrix or data frame used for automatic nuisance
cross-fitting.}

\item{p}{Treatment probability used in residualization. The default is the
empirical treatment proportion. Supply the known assignment probability
when available.}

\item{n_folds}{Number of nuisance cross-fitting folds. Default is 5.}

\item{nuisance_learner}{Either \code{"SuperLearner"} or a custom prediction
function; see \code{\link[=residualize_outcome]{residualize_outcome()}}.}

\item{SL.library}{SuperLearner library used for automatic residualization.
The default includes mean, GLM, GAM, glmnet, neural-network, earth, and
pruned-tree learners.}

\item{SL_folds}{Number of inner SuperLearner folds. Default is 5.}

\item{learner_args}{Additional arguments for the nuisance learner.}
}
\value{
A list that contains the following items: \item{stat}{The estimated
statistic for the test of heterogeneity.} \item{pval}{The p-value of the null
hypothesis (that the treatment effects are homogeneous)}
}
\description{
This function calculates statistics related to the test of heterogeneous treatment effects across groups.
}
\details{
The details of the methods for this design are given in Imai and Li (2025).
The supplied outcome is centered by default. Optionally, arm-specific
outcome regressions can be used to construct a cross-fitted residualized
outcome before conducting the test.
}
\examples{
T <- rep(c(1, 0), 20)
tau <- seq_len(40) / 40
Y <- tau * T + rnorm(40)
hettestlist <- het.test(T,tau,Y,ngates=5)
hettestlist$stat
hettestlist$pval
}
\references{
Imai and Li (2025). \dQuote{Statistical Inference for Heterogeneous Treatment Effects Discovered by Generic Machine Learning in Randomized Experiments},
}
\author{
Michael Lingzhi Li, Technology and Operations Management, Harvard Business School
\email{mili@hbs.edu}, \url{https://www.michaellz.com/};
}
\keyword{evaluation}
