% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_plot.R
\name{plot.hte}
\alias{plot.hte}
\title{Plot the GATE estimate}
\usage{
\method{plot}{hte}(x, ...)
}
\arguments{
\item{x}{An table object. This is typically an output of \code{evaluate_hte()} function.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the GATE estimate
}
