% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_plot.R
\name{plot_CI.hte}
\alias{plot_CI.hte}
\title{Plot the uniform confidence interval}
\usage{
plot_CI.hte(x, ...)
}
\arguments{
\item{x}{An object of \code{evaluate_hte()} class. This is typically an output of \code{evaluate_hte()} function.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the uniform confidence interval
}
