% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_summary.R
\name{print.summary.hte}
\alias{print.summary.hte}
\title{Print}
\usage{
\method{print}{summary.hte}(x, ...)
}
\arguments{
\item{x}{An object of \code{summary.hte} class. This is typically an output of \code{summary.hte()} function.}

\item{...}{Other parameters. Currently not supported.}
}
\value{
No return value, called for side effects (prints summary tables to console).
}
\description{
Print
}
