% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_summary.R
\name{print.summary.test_hte}
\alias{print.summary.test_hte}
\title{Print}
\usage{
\method{print}{summary.test_hte}(x, ...)
}
\arguments{
\item{x}{An object of \code{summary.test_hte} class. This is typically an output of \code{summary.test_hte()} function.}

\item{...}{Other parameters.}
}
\value{
No return value, called for side effects (prints test results to console).
}
\description{
Print
}
