% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_summary.R
\name{summary.hte}
\alias{summary.hte}
\title{Summarize Heterogeneity and Consistency Tests}
\usage{
\method{summary}{hte}(object, ...)
}
\arguments{
\item{object}{An object of \code{evaluate_hte} class (typically an output of \code{evaluate_hte()} function).}

\item{...}{Other parameters.}
}
\description{
Summarize Heterogeneity and Consistency Tests
}
