% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{evaluate_hte}
\alias{evaluate_hte}
\title{Evaluate Heterogeneous Treatment Effects}
\usage{
evaluate_hte(fit, ...)
}
\arguments{
\item{fit}{Fitted model. Usually an output from \code{estimate_hte}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
An object of \code{hte} class
}
\description{
Evaluate Heterogeneous Treatment Effects
}
