% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_plot.R
\name{plot_CI.hte}
\alias{plot_CI.hte}
\title{Plot the uniform confidence interval}
\usage{
\method{plot_CI}{hte}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{An object of \code{evaluate_hte()} class. This is typically an output of \code{evaluate_hte()} function.}

\item{alpha}{Significance level. Default is 0.05.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the uniform confidence interval
}
