% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte_plot.R
\name{plot_CI}
\alias{plot_CI}
\title{Plot Confidence Intervals}
\usage{
plot_CI(x, ...)
}
\arguments{
\item{x}{An object for which a plot is desired.}

\item{...}{Further arguments passed to methods.}
}
\value{
A ggplot2 object displaying uniform and pointwise confidence intervals for heterogeneous treatment effects.
}
\description{
A generic function to plot uniform and pointwise confidence intervals for HTE objects.
}
